Turned handlePlease into a coroutine as required by discord.tcl 0.5.0.
This is to facilitate getting the response from a HTTP request in an
event-driven manner. A coroutine context for handlePlease will yield after
calling discord::sendMessage, and will be resumed by a HTTP callback. The name
returned from sendMessage is a coroutine which can be called once to get a list
of data and httpCode.

Other changes include adding a tclqBot logger, prepending ':' to variables
inside sqlite3 evals to avoid SQL injection.